Correct stack identity when delete nested
1. Update an aws asg with invalid 'VPCZoneIdentifier',
the stack update failed.
2. Update again, the stack is always in 'UPDATE_IN_PROGRESS'.

We remove the backup resources first when updating,
but the backup of stack_resource can't be deleted,
due we set an incorrect name of stack_identity before
rpc calling 'delete_stack', due the name is not match the
nested_stack's name.

Closes-Bug: #1478480
(cherry picked from commit ab234eb65871be3607a5702d6a01a6be5f6b59d5)

+++ heat/engine/resources/stack_resource.py
@@ -398,10 +398,15 @@ class StackResource(resource.Resource):
         '''
         Delete the nested stack.
         '''
-        stack_identity = identifier.HeatIdentifier(
-            self.context.tenant_id,
-            self.physical_resource_name(),
-            self.resource_id)
+        try:
+            stack = self.nested()
+        except exception.NotFound:
+            return
+
+        if stack is None:
+            return
+
+        stack_identity = stack.identifier()
 
         try:
             self.rpc_client().delete_stack(self.context, stack_identity)
